+++ inst/Rmd_vignette/style.R
@@ -1,6 +1,0 @@
-options(rstudio.markdownToHTML = 
-  function(inputFile, outputFile) {      
-    require(markdown)
-    markdownToHTML(inputFile, outputFile, stylesheet=file.path(getwd(), 'css/style.css'))
-  }
-)